+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `SUBMISSION_TIMESTAMP` type to workflow event and approval step conditional predicates schema
+
 ## [13.2.1] - 2026-07-23
 
 ### Added